+++ src/index.js
@@ -1,5 +1,5 @@
 (function(){
-  var v = function(selector, history) {
+  var v = function(selector, context={methodHistory: [], selectorHistory: []}) {
     if (!(this instanceof v)) {
       return new v(selector);
     } 
@@ -12,12 +12,16 @@
         this.selector = selector;
       }
     }
-    if (!history) {
+    this.version = '4.6.0';
+    this.methodHistory = context.methodHistory;
+
+    if (context.selectorHistory.length === 0) {
       this.selectorHistory = [];
       this.selectorHistory.push(selector);
     } else {
-      this.selectorHistory = history;
+      this.selectorHistory = context.selectorHistory;
     }
+    
     // Utility functions
     var error = (msg, selector, method, parameters)=>{
       var errMsg = (msg)=>{
@@ -25,10 +29,10 @@
       };
       if (msg === 'undefinedNode') {
         errMsg('Selector is not a valid node.');
-      } else if (msg === 'notFunc') {
-        errMsg(`Parameter passed to the ready method is not of the type 'function'.`);
+      } else if (msg === 'notType') {
+        errMsg(`Parameter passed to the ${method} method is not of the type '${parameters}'.`);
       } else {
-        errMsg();
+        errMsg(msg);
       }
     };
     var isElement = (element)=>{
@@ -60,11 +64,15 @@
     v.prototype.typeOf = (input)=>{
       return Object.prototype.toString.call(input).replace(/^\[object (.+)\]$/, '$1').toLowerCase();
     };
-    v.prototype.handler = (data)=>{
-      // If the selector is updated, start a new instance with the updated selectoreter.
+    v.prototype.handler = (data, opts)=>{
+      // If the selector is updated, start a new instance with the updated selector.
       let _selector = data ? data : this.selector ? this.selector : selector;
       this.selectorHistory.push(_selector);
-      return new v(_selector, this.selectorHistory);
+      this.methodHistory.push(opts.method);
+      return new v(_selector, {
+        selectorHistory: this.selectorHistory,
+        methodHistory: this.methodHistory
+      });
     };
     v.prototype.move = (fromIndex, toIndex)=>{
       try {
@@ -81,7 +89,7 @@
         return uniq;
       } else {
         this.selector = uniq;
-        return this.handler();
+        return this.handler(null, {method: 'uniq'});
       }
     };
     v.prototype.slice = (nodeList)=>{
@@ -89,7 +97,7 @@
     };
     var assignNodes = (nodes)=>{
       this.nodes = this.slice(nodes);
-      this.node = this.nodes[0];
+      this.node = this.nodes.length > 0 ? this.nodes[0] : null;
     };
     // Turn the CSS selector into a node, pass an existing node to this.nodes, which is used by all methods.
     v.prototype.query = (el, _selector)=>{
@@ -116,7 +124,7 @@
           }
 
         }
-        if (typeof this.node === 'undefined') {
+        if (!this.node) {
           this.nonElement = selector;
         }
       } else {
@@ -143,7 +151,7 @@
       Promise._setImmediateFn(setAsap);
       return new Promise((resolve, reject)=>{
         var _resolve = (data)=>{
-          let _data = options && options.chain ? this.handler(data) : data;
+          let _data = options && options.chain ? this.handler(data, {method: 'ajax'}) : data;
           if (typeof _data !== 'undefined' && _data) {
             resolve(_data);
           }
@@ -184,34 +192,34 @@
     // v(selector).get(0) -> <div></div>
     v.prototype.get = (i)=>{
       this.selector = this.nodes[i];
-      return this.handler();
+      return this.handler(null, {method: 'get'});
     };
     // Event methods
     v.prototype.ready = (func)=>{
       if (func && typeof func !== 'undefined' && typeof func === 'function') {
         document.addEventListener('DOMContentLoaded', func);
       } else {
-        error('notFunc', '', 'ready', 'function');
+        error('notType', '', 'ready', 'function');
       }
     };
     v.prototype.load = (func)=>{
       if (func && typeof func !== 'undefined' && typeof func === 'function') {
         document.addEventListener('load', func);
       } else {
-        error('notFunc', '', 'load', 'function');
+        error('notType', '', 'load', 'function');
       }
     };
     v.prototype.on = (event, func)=>{
       this.for(this.nodes, (i)=>{
         i.addEventListener(event, func);
       });
-      return this.handler();
+      return this.handler(null, {method: 'on'});
     };
     v.prototype.off = (event, func)=>{
       this.for(this.nodes, (i)=>{
         i.removeEventListener(event, func);
       });
-      return this.handler();
+      return this.handler(null, {method: 'off'});
     };
     v.prototype.trigger = (event)=>{
       if (this.node.fireEvent) {
@@ -221,7 +229,7 @@
         evObj.initEvent(event, true, false);
         this.node.dispatchEvent(evObj);
       }
-      return this.handler();
+      return this.handler(null, {method: 'trigger'});
     };
     v.prototype.click = (func)=>{
       this.for(this.nodes, (i)=>{
@@ -231,20 +239,20 @@
           this.trigger('click');
         }
       });
-      return this.handler();
+      return this.handler(null, {method: 'click'});
     };
     // DOM traversal and manipulation methods
     v.prototype.filter = (func)=>{
       Array.prototype.filter.call(this.nodes, func);
-      return this.handler();
+      return this.handler(null, {method: 'filter'});
     };
     v.prototype.each = (func)=>{
       Array.prototype.forEach.call(this.nodes, func);
-      return this.handler();
+      return this.handler(null, {method: 'each'});
     };
     v.prototype.map = (func)=>{
       Array.prototype.map.call(this.nodes, func);
-      return this.handler();
+      return this.handler(null, {method: 'map'});
     };
     v.prototype.find = (_selector)=>{
       if (!isElement(_selector)) {
@@ -265,50 +273,50 @@
       } else {
         this.selector = _selector;
       }
-      return this.handler();
+      return this.handler(null, {method: 'find'});
     };
     v.prototype.end = ()=>{
       this.selector = this.selectorHistory[0];
-      return this.handler();
+      return this.handler(null, {method: 'end'});
     };
     v.prototype.hide = ()=>{
       this.for(this.nodes, (i)=>{
         i.style.display = 'none';
       });
-      return this.handler();
+      return this.handler(null, {method: 'hide'});
     };
     v.prototype.show = ()=>{
       this.for(this.nodes, (i)=>{
         i.style.display = 'block';
       });
-      return this.handler();
+      return this.handler(null, {method: 'show'});
     };
     v.prototype.remove = ()=>{
       this.for(this.nodes, (i)=>{
         i.parentNode.removeChild(i);
       });
-      return this.handler();
+      return this.handler(null, {method: 'remove'});
     };
     v.prototype.empty = ()=>{
       this.for(this.nodes, (i)=>{
         i.innerHTML = '';
       });
-      return this.handler();
+      return this.handler(null, {method: 'empty'});
     };
     v.prototype.clone = ()=>{
       var clone = this.node.cloneNode(true);
       this.selector = clone;
-      return this.handler();
+      return this.handler(null, {method: 'clone'});
     };
     v.prototype.wrap = (tag)=>{
       this.for(this.nodes, (i)=>{
         i.outerHTML = `${tag.match(/<(.|\n)*?>/g)[0]}${i.outerHTML}`;
       });
-      return this.handler();
+      return this.handler(null, {method: 'wrap'});
     };
     v.prototype.parent = ()=>{
       this.selector = this.node.parentNode;
-      return this.handler();
+      return this.handler(null, {method: 'parent'});
     };
     v.prototype.parents = (el)=>{
       let parent = this.node.parentNode;
@@ -330,7 +338,7 @@
       } else {
         this.selector = _parents;
       }
-      return this.handler();
+      return this.handler(null, {method: 'parents'});
     };
     v.prototype.children = (el)=>{
       let children = this.slice(this.nodes[0].children);
@@ -347,7 +355,7 @@
       } else {
         this.selector = children;
       }
-      return this.handler();
+      return this.handler(null, {method: 'children'});
     };
     v.prototype.allChildren = (_el)=> {
       var __el = _el ? this.slice(this.query(this.node, _el))[0] : this.node;
@@ -364,7 +372,7 @@
       };
       recurse(__el);
       this.selector = arr;
-      return this.handler();
+      return this.handler(null, {method: 'allChildren'});
     };
     v.prototype.isEmpty = ()=>{
       return !this.node.hasChildNodes();
@@ -377,11 +385,11 @@
     };
     v.prototype.next = ()=>{
       this.selector = this.node.nextElementSibling;
-      return this.handler();
+      return this.handler(null, {method: 'next'});
     };
     v.prototype.prev = ()=>{
       this.selector = this.node.previousElementSibling;
-      return this.handler();
+      return this.handler(null, {method: 'prev'});
     };
     v.prototype.addClass = (_class)=>{
       var classArr = _class.split(' ');
@@ -390,7 +398,7 @@
           i.classList.add(y);
         });
       });
-      return this.handler();
+      return this.handler(null, {method: 'addClass'});
     };
     v.prototype.removeClass = (_class)=>{
       var classArr = _class.split(' ');
@@ -399,7 +407,7 @@
           i.classList.remove(y);
         });
       });
-      return this.handler();
+      return this.handler(null, {method: 'removeClass'});
     };
     v.prototype.toggleClass = (_class)=>{
       var classArr = _class.split(' ');
@@ -408,7 +416,7 @@
           i.classList.toggle(y);
         });
       });
-      return this.handler();
+      return this.handler(null, {method: 'toggleClass'});
     };
     v.prototype.hasClass = (_class)=>{
       var bool = this.node.classList.contains(_class);
@@ -418,7 +426,7 @@
       this.for(this.nodes, (i)=>{
         i.removeAttribute(attr);
       });
-      return this.handler();
+      return this.handler(null, {method: 'removeAttr'});
     };
     // v(selector).attr() returns an object of camelized attribute keys. 
     // v(selector).attr({dataId: '0'}) -> <div data-id="0"></div>
@@ -428,14 +436,14 @@
         if (props) {
           if (props2 && this.typeOf(props2) === 'string') {
             i.setAttribute(this.decamelize(props), props2);
-            _return = this.handler();
+            _return = this.handler(null, {method: 'attr'});
           } else {
             this.forIn(props, (y)=>{
               if (this.typeOf(props) === 'string') {
                 _return = i.attributes[props].value;
               } else {
                 i.setAttribute(this.decamelize(y), props[y]);
-                _return = this.handler();
+                _return = this.handler(null, {method: 'attr'});
               }
             });
           }
@@ -461,7 +469,7 @@
             i.style[y] = props[y];
           });
         });
-        return this.handler();
+        return this.handler(null, {method: 'css'});
       } else {
         return getComputedStyle(this.node);
       }
@@ -471,7 +479,7 @@
         this.for(this.nodes, (i)=>{
           i.value = string;
         });
-        return this.handler();
+        return this.handler(null, {method: 'val'});
       } else {
         return this.node.value;
       }
@@ -522,25 +530,24 @@
           i.innerHTML = contents;
         }
       });
-      return contents ? this.handler() : output;
+      return contents ? this.handler(null, {method: 'html'}) : output;
     };
     v.prototype.json = (input)=>{
-      let _input = this.nodes ? this.nodes : input;
-      let output = JSON.stringify(_input);
-      if (output && typeof _input !== 'string') {
+      let _input = this.nodes ? this.nodes : this.nonElement ? this.nonElement : input;
+      try {
         return JSON.stringify(_input);
-      } else {
-        error(`Parameter passed to the json method cannot be converted into valid JSON.`);
+      } catch (e) {
+        error(e, '', 'json', 'serializable input');
       }
     };
     v.prototype.parseJSON = (string)=>{
       string = this.nonElement ? this.nonElement : string;
       try {
         var output = JSON.parse(string);
+        return output;
       } catch (e) {
-        error(`Parameter passed to the parseJSON method is not valid JSON.\n${e}`);
+        error(e, '', 'parseJSON', 'valid JSON');
       }
-      return output;
     };
     v.prototype.type = (input)=>{
       input = queryParameter(input, true);
@@ -550,9 +557,9 @@
       this.for(this.nodes, (i)=>{
         if (string && typeof string === 'string') {
           i.outerHTML = string;
-          return this.handler();
+          return this.handler(null, {method: 'replaceWith'});
         } else {
-          error(`Parameter passed to the replaceWith method is not of the type 'string'.`);
+          error('notType', '', 'replaceWith', 'string');
         }
       });
     };
@@ -565,39 +572,39 @@
           i.textContent = contents;
         }
       });
-      return contents ? this.handler() : output;
+      return contents ? this.handler(null, {method: 'text'}) : output;
     };
     v.prototype.insertBefore = (el1, el2)=>{
       this.for(this.nodes, (i)=>{
         i.insertBefore(el1, el2);
       });
-      return this.handler();
+      return this.handler(null, {method: 'insertBefore'});
     };
     v.prototype.prepend = (el)=>{ 
       let _element = queryElement(el);
       this.for(this.nodes, (i)=>{
         i.insertBefore(_element, i.firstChild);
       });
-      return this.handler();
+      return this.handler(null, {method: 'prepend'});
     };
     v.prototype.append = (el)=>{ 
       let _element = queryElement(el);
       this.for(this.nodes, (i)=>{
         i.appendChild(_element);
       });
-      return this.handler();
+      return this.handler(null, {method: 'append'});
     };
     v.prototype.after = (string)=>{
       this.for(this.nodes, (i)=>{
         i.insertAdjacentHTML('afterend', string);
       });
-      return this.handler();
+      return this.handler(null, {method: 'after'});
     };
     v.prototype.before = (string)=>{
       this.for(this.nodes, (i)=>{
         i.insertAdjacentHTML('beforebegin', string);
       });
-      return this.handler();
+      return this.handler(null, {method: 'before'});
     };
     v.prototype.contains = (text)=>{
       let textContent = null;
@@ -643,10 +650,7 @@
     };
     v.prototype.decamelize = (string)=>{
       string = queryParameter(string, false);
-      return string
-        .replace(/([a-z])([A-Z])/g, '$1-$2')
-        .replace(/\b([A-Z]+)([A-Z])([a-z])/, '$1-$2$3')
-        .replace(/^./, function(str){ return str.toLowerCase(); });
+      return string.replace(/([a-z])([A-Z])/g, '$1-$2').replace(/\b([A-Z]+)([A-Z])([a-z])/, '$1-$2$3').replace(/^./, function(str){ return str.toLowerCase(); });
     };
     v.prototype = {
       get n(){